教えてもらった部分を修正 (テストはコケる)
