--- lib/dev/api.d.ts
@@ -1,16 +1,0 @@
-export interface ZipCloudJson {
-  message: null,
-  results: [
-    {
-      address1: string,
-      address2: string,
-      address3: string,
-      kana1: string,
-      kana2: string,
-      kana3: string,
-      prefcode: string,
-      zipcode: string
-    }
-  ],
-  status: number
-}
